Moved Drug and Pharmacy classes to own folder in models directory
Moved Drug and Pharmacy classes to own folder in models directory
