+++ models/drug/index.js
@@ -1,0 +1,1 @@
+export { Drug } from "./drug"
